[libc] Move mbtowc, mbstowcs and inverse functions to stdlib.h
These functions should be declared in stdlib.h, not wchar.h, as confusing as it is.
Move them to the proper header file and matching directories in src/ and test/
trees.

This was discovered while testing libc++ build against llvm-libc, which
re-declares functions like mbtowc in std-namespace in <cstdlib> header,
and then uses those functions in its locale implementation.
