+++ libc/src/stdlib/mbtowc.cpp
@@ -1,4 +1,4 @@
-//===-- Implementation of mbstowcs ----------------------------------------===//
+//===-- Implementation of mbtowc -----------------------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,35 +6,32 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "src/wchar/mbstowcs.h"
+#include "src/stdlib/mbtowc.h"
 
 #include "hdr/types/size_t.h"
 #include "hdr/types/wchar_t.h"
 #include "src/__support/common.h"
 #include "src/__support/libc_errno.h"
 #include "src/__support/macros/config.h"
-#include "src/__support/macros/null_check.h"
-#include "src/__support/wchar/mbsnrtowcs.h"
+#include "src/__support/wchar/mbrtowc.h"
 #include "src/__support/wchar/mbstate.h"
 
 namespace LIBC_NAMESPACE_DECL {
 
-LLVM_LIBC_FUNCTION(size_t, mbstowcs,
-                   (wchar_t *__restrict pwcs, const char *__restrict s,
+LLVM_LIBC_FUNCTION(int, mbtowc,
+                   (wchar_t *__restrict pwc, const char *__restrict s,
                     size_t n)) {
-  LIBC_CRASH_ON_NULLPTR(s);
-  // If destination is null, ignore n
-  n = pwcs == nullptr ? SIZE_MAX : n;
-  static internal::mbstate internal_mbstate;
-  const char *temp = s;
-  auto ret = internal::mbsnrtowcs(pwcs, &temp, SIZE_MAX, n, &internal_mbstate);
-
-  if (!ret.has_value()) {
+  // returns 0 since UTF-8 encoding is not state-dependent
+  if (s == nullptr)
+    return 0;
+  internal::mbstate internal_mbstate;
+  auto ret = internal::mbrtowc(pwc, s, n, &internal_mbstate);
+  if (!ret.has_value() || static_cast<int>(ret.value()) == -2) {
     // Encoding failure
-    libc_errno = ret.error();
+    libc_errno = EILSEQ;
     return -1;
   }
-  return ret.value();
+  return static_cast<int>(ret.value());
 }
 
 } // namespace LIBC_NAMESPACE_DECL
