+++ libc/src/stdlib/wcstombs.h
@@ -1,4 +1,4 @@
-//===-- Implementation header for wctomb ------------------------*- C++ -*-===//
+//===-- Implementation header for wcstombs --------------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,17 +6,17 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef LLVM_LIBC_SRC_WCHAR_WCTOMB_H
-#define LLVM_LIBC_SRC_WCHAR_WCTOMB_H
+#ifndef LLVM_LIBC_SRC_STDLIB_WCSTOMBS_H
+#define LLVM_LIBC_SRC_STDLIB_WCSTOMBS_H
 
-#include "hdr/types/mbstate_t.h"
+#include "hdr/types/size_t.h"
 #include "hdr/types/wchar_t.h"
 #include "src/__support/macros/config.h"
 
 namespace LIBC_NAMESPACE_DECL {
 
-int wctomb(char *s, wchar_t wc);
+size_t wcstombs(char *__restrict s, const wchar_t *__restrict pwcs, size_t n);
 
 } // namespace LIBC_NAMESPACE_DECL
 
-#endif // LLVM_LIBC_SRC_WCHAR_WCTOMB_H
+#endif // LLVM_LIBC_SRC_STDLIB_WCSTOMBS_H
